hr-chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
databricks-sql-connector[sqlalchemy]
databricks-sqlalchemy
langchain
langchain_classic
langchain-core
langchain_community
langchain_openai
langgraph
openai
